\name{lmpVal}
\alias{lmpVal}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Linear Model P-Val
}
\description{
Returns a lienar model's p val
}
\usage{
lmpVal(modelObject)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{modelObject}{
Object of class 'lm'
}
}

