\name{readMetadata}
\alias{readMetadata}
\title{
readMetadata
}
\description{
Simply reads in a CSV and converts the treatments column to a vector
}
\usage{
readMetadata(fileName)
}
\arguments{
  \item{fileName}{
  Exact location of metadata csv
}
}
\examples{

readMetadata("/Users/Path/Resources/info.csv")

